requirements.txt
requests
python-dotenv
Pillow
svgwrite